move part cutlist generation to new core utils file
