+++ plugins/plugin-title/src/index.js
@@ -1,4 +1,5 @@
 import pkg from '../package.json'
+import {getPartCutlist} from '@freesewing/core'
 
 const style = `
 text.title-nr {
@@ -84,7 +85,7 @@ export default {
         shift += 8
       }
       if (so.cutList) {
-        let cutList = this.context.config.cutList?.[(this.name || so.title || '').replace(/_cutPiece\d+/, '')]
+        let cutList = getPartCutlist((this.name || so.title || '').replace(/_cutPiece\d+/, ''), this.context.config)
         let cutCount = 1;
         let pairs = '';
         if (cutList) {
